Implement comprehensive SEO optimization and Vercel Analytics for Kenya market dominance
