+++ app/layout.tsx
@@ -9,16 +9,26 @@ import { Toaster } from "@/components/ui/toaster";
 import { AuthProvider } from "@/contexts/AuthContext";
 import { MarketplaceProvider } from "@/contexts/MarketplaceContext";
 import { AdminProvider } from "@/contexts/AdminContext";
+import { Analytics } from "@vercel/analytics/react";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "JABA Automobiles",
-  description: "Your trusted partner for all automotive needs",
-  // Ensure absolute URLs for social previews
+  title: {
+    default: "JABA Automobiles - Premium Vehicle Marketplace",
+    template: "%s | JABA Automobiles",
+  },
+  description: "Buy and sell premium vehicles at JABA Automobiles. Browse our extensive collection of cars, SUVs, motorbikes, and luxury vehicles. Direct imports, second-hand, and certified pre-owned vehicles available.",
+  keywords: "cars for sale, vehicle marketplace, buy cars, sell cars, SUV, motorbikes, luxury cars, direct import, second hand cars, JABA Automobiles, Kenya cars",
+  authors: [{ name: "JABA Automobiles" }],
+  creator: "JABA Automobiles",
+  publisher: "JABA Automobiles",
   metadataBase: new URL(
-    process.env.NEXT_PUBLIC_SITE_URL || "https://example.com"
+    process.env.NEXT_PUBLIC_SITE_URL || "https://jabaautomotives.com"
   ),
+  alternates: {
+    canonical: "/",
+  },
   icons: {
     icon: "/logo.png",
     shortcut: "/logo.png",
@@ -26,23 +36,41 @@ export const metadata: Metadata = {
   },
   openGraph: {
     type: "website",
-    title: "JABA Automobiles",
-    description: "Your trusted partner for all automotive needs",
+    locale: "en_US",
     url: "/",
+    siteName: "JABA Automobiles",
+    title: "JABA Automobiles - Premium Vehicle Marketplace",
+    description: "Buy and sell premium vehicles at JABA Automobiles. Browse our extensive collection of cars, SUVs, motorbikes, and luxury vehicles.",
     images: [
       {
         url: "/logo.png",
-        width: 512,
-        height: 512,
-        alt: "JABA Automobiles",
+        width: 1200,
+        height: 630,
+        alt: "JABA Automobiles - Premium Vehicle Marketplace",
       },
     ],
   },
   twitter: {
-    card: "summary",
-    title: "JABA Automobiles",
-    description: "Your trusted partner for all automotive needs",
+    card: "summary_large_image",
+    site: "@jabaautomotives",
+    creator: "@jabaautomotives",
+    title: "JABA Automobiles - Premium Vehicle Marketplace",
+    description: "Buy and sell premium vehicles at JABA Automobiles. Browse our extensive collection of cars, SUVs, motorbikes, and luxury vehicles.",
     images: ["/logo.png"],
+  },
+  robots: {
+    index: true,
+    follow: true,
+    googleBot: {
+      index: true,
+      follow: true,
+      "max-video-preview": -1,
+      "max-image-preview": "large",
+      "max-snippet": -1,
+    },
+  },
+  verification: {
+    google: process.env.GOOGLE_SITE_VERIFICATION,
   },
 };
 
@@ -53,6 +81,12 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en" suppressHydrationWarning>
+      <head>
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <meta name="theme-color" content="#000000" />
+      </head>
       <body className={inter.className}>
         <ThemeProvider
           attribute="class"
@@ -68,6 +102,7 @@ export default function RootLayout({
                   <Footer />
                 </div>
                 <Toaster />
+                <Analytics />
               </MarketplaceProvider>
             </AdminProvider>
           </AuthProvider>
