+++ app/sitemap.ts
@@ -1,0 +1,92 @@
+import { MetadataRoute } from 'next'
+import { collection, getDocs, query, where } from 'firebase/firestore'
+import { db } from '@/lib/firebase'
+
+export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
+  const siteUrl = process.env.NEXT_PUBLIC_SITE_URL || 'https://jabaautomotives.com'
+
+  // Static pages
+  const staticPages = [
+    {
+      url: siteUrl,
+      lastModified: new Date(),
+      changeFrequency: 'daily' as const,
+      priority: 1,
+    },
+    {
+      url: `${siteUrl}/marketplace`,
+      lastModified: new Date(),
+      changeFrequency: 'hourly' as const,
+      priority: 0.9,
+    },
+    {
+      url: `${siteUrl}/second-hand`,
+      lastModified: new Date(),
+      changeFrequency: 'daily' as const,
+      priority: 0.8,
+    },
+    {
+      url: `${siteUrl}/direct-import`,
+      lastModified: new Date(),
+      changeFrequency: 'daily' as const,
+      priority: 0.8,
+    },
+    {
+      url: `${siteUrl}/motorbikes`,
+      lastModified: new Date(),
+      changeFrequency: 'daily' as const,
+      priority: 0.7,
+    },
+    {
+      url: `${siteUrl}/about`,
+      lastModified: new Date(),
+      changeFrequency: 'monthly' as const,
+      priority: 0.6,
+    },
+    {
+      url: `${siteUrl}/contact`,
+      lastModified: new Date(),
+      changeFrequency: 'monthly' as const,
+      priority: 0.6,
+    },
+    {
+      url: `${siteUrl}/financing`,
+      lastModified: new Date(),
+      changeFrequency: 'monthly' as const,
+      priority: 0.7,
+    },
+    {
+      url: `${siteUrl}/insurance`,
+      lastModified: new Date(),
+      changeFrequency: 'monthly' as const,
+      priority: 0.7,
+    },
+  ]
+
+  // Dynamic vehicle pages
+  let vehiclePages: MetadataRoute.Sitemap = []
+  
+  try {
+    // Get all active vehicle listings
+    const vehiclesQuery = query(
+      collection(db, 'vehicleListings'),
+      where('status', '==', 'active')
+    )
+    
+    const vehiclesSnapshot = await getDocs(vehiclesQuery)
+    
+    vehiclePages = vehiclesSnapshot.docs.map((doc) => {
+      const data = doc.data()
+      return {
+        url: `${siteUrl}/vehicles/${doc.id}`,
+        lastModified: data.updatedAt?.toDate() || data.createdAt?.toDate() || new Date(),
+        changeFrequency: 'weekly' as const,
+        priority: 0.8,
+      }
+    })
+  } catch (error) {
+    console.error('Error generating vehicle sitemap entries:', error)
+  }
+
+  return [...staticPages, ...vehiclePages]
+}
