+++ app/vehicles/[id]/loading.tsx
@@ -1,0 +1,99 @@
+import { Skeleton } from "@/components/ui/skeleton"
+
+export default function VehicleLoading() {
+  return (
+    <div className="container mx-auto px-4 py-8">
+      {/* Back button skeleton */}
+      <Skeleton className="h-6 w-40 mb-4" />
+
+      <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 mt-4">
+        {/* Image gallery skeleton */}
+        <div className="space-y-4">
+          <Skeleton className="h-96 w-full rounded-lg" />
+          <div className="flex gap-2">
+            <Skeleton className="h-20 w-20 rounded" />
+            <Skeleton className="h-20 w-20 rounded" />
+            <Skeleton className="h-20 w-20 rounded" />
+            <Skeleton className="h-20 w-20 rounded" />
+          </div>
+        </div>
+
+        {/* Vehicle details skeleton */}
+        <div className="space-y-6">
+          <div className="flex items-start justify-between">
+            <div className="space-y-2">
+              <Skeleton className="h-8 w-80" />
+              <Skeleton className="h-6 w-32" />
+            </div>
+            <Skeleton className="h-6 w-20" />
+          </div>
+
+          <Skeleton className="h-20 w-full" />
+
+          {/* Features skeleton */}
+          <div className="space-y-2">
+            <Skeleton className="h-4 w-16" />
+            <div className="flex flex-wrap gap-2">
+              <Skeleton className="h-6 w-20" />
+              <Skeleton className="h-6 w-24" />
+              <Skeleton className="h-6 w-18" />
+              <Skeleton className="h-6 w-22" />
+            </div>
+          </div>
+
+          {/* Action buttons skeleton */}
+          <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
+            <Skeleton className="h-10 w-full" />
+            <Skeleton className="h-10 w-full" />
+          </div>
+
+          <div className="flex gap-4">
+            <Skeleton className="h-10 flex-1" />
+            <Skeleton className="h-10 w-20" />
+          </div>
+
+          <Skeleton className="h-10 w-full" />
+
+          {/* Contact info skeleton */}
+          <div className="space-y-3">
+            <div className="flex items-center gap-3">
+              <Skeleton className="h-10 w-10 rounded-full" />
+              <div className="space-y-1">
+                <Skeleton className="h-4 w-32" />
+                <Skeleton className="h-3 w-48" />
+              </div>
+            </div>
+            <Skeleton className="h-3 w-28" />
+          </div>
+
+          {/* Specs skeleton */}
+          <div className="space-y-3">
+            <Skeleton className="h-5 w-24" />
+            <div className="grid grid-cols-2 gap-4">
+              {Array.from({ length: 8 }).map((_, i) => (
+                <div key={i} className="space-y-1">
+                  <Skeleton className="h-3 w-20" />
+                  <Skeleton className="h-4 w-16" />
+                </div>
+              ))}
+            </div>
+          </div>
+        </div>
+      </div>
+
+      {/* Similar vehicles skeleton */}
+      <div className="mt-16 space-y-4">
+        <Skeleton className="h-6 w-40" />
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+          {Array.from({ length: 3 }).map((_, i) => (
+            <div key={i} className="space-y-3">
+              <Skeleton className="h-48 w-full rounded-lg" />
+              <Skeleton className="h-5 w-3/4" />
+              <Skeleton className="h-4 w-1/2" />
+            </div>
+          ))}
+        </div>
+      </div>
+    </div>
+  )
+}
