+++ components/VehicleGrid.tsx
@@ -1,56 +1,63 @@
 "use client"
 
-import { useState } from "react"
 import Link from "next/link"
 import { Card, CardContent, CardFooter } from "@/components/ui/card"
 import { Button } from "@/components/ui/button"
-import CarListingSchema from "./CarListingSchema"
-import SocialShare from "./SocialShare"
+import { Skeleton } from "@/components/ui/skeleton"
+import { VehicleListing } from "@/contexts/MarketplaceContext"
 import LazyImage from "./LazyImage"
 
-// ... (Vehicles array remains the same)
+interface VehicleGridProps {
+  vehicles: VehicleListing[]
+  loading: boolean
+}
 
-export default function VehicleGrid() {
-  const [vehicles] = useState(Vehicles)
+export function VehicleGrid({ vehicles, loading }: VehicleGridProps) {
+  if (loading) {
+    return (
+      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+        {Array.from({ length: 6 }).map((_, i) => (
+          <Card key={i} className="overflow-hidden">
+            <Skeleton className="h-48 w-full" />
+            <CardContent className="p-4">
+              <Skeleton className="h-6 w-3/4 mb-2" />
+              <Skeleton className="h-4 w-1/2 mb-2" />
+              <Skeleton className="h-4 w-1/3" />
+            </CardContent>
+            <CardFooter>
+              <Skeleton className="h-10 w-full" />
+            </CardFooter>
+          </Card>
+        ))}
+      </div>
+    )
+  }
 
   return (
     <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
       {vehicles.map((vehicle) => (
         <Card key={vehicle.id} className="overflow-hidden">
-          <CarListingSchema
-            name={vehicle.name}
-            description={` ${vehicle.brand} ${vehicle.model} available for purchase with cryptocurrency`}
-            brand={vehicle.brand}
-            model={vehicle.model}
-            year={vehicle.year}
-            price={vehicle.price}
-            currency={vehicle.currency}
-            image={`https://cryptodrive-.com${vehicle.image}`}
-            url={`https://cryptodrive-.com/vehicles/${vehicle.id}`}
-          />
           <div className="relative h-48">
             <LazyImage
-              src={vehicle.image}
-              alt={`${vehicle.name} -  car available for purchase with cryptocurrency`}
+              src={vehicle.images?.[0] || "/placeholder.jpg"}
+              alt={`${vehicle.year} ${vehicle.make} ${vehicle.model}`}
               width={400}
               height={300}
             />
           </div>
           <CardContent className="p-4">
-            <h3 className="text-xl font-bold mb-2">{vehicle.name}</h3>
+            <h3 className="text-xl font-bold mb-2">{vehicle.title}</h3>
             <p className="text-muted-foreground mb-2">
-              {vehicle.price.toLocaleString()} {vehicle.currency}
+              {vehicle.year} {vehicle.make} {vehicle.model}
             </p>
-            <p className="text-primary font-semibold">{vehicle.cryptoPrice}</p>
+            <p className="text-primary font-semibold">
+              {typeof vehicle.price === 'string' ? vehicle.price : `KES ${vehicle.price?.toLocaleString()}`}
+            </p>
           </CardContent>
-          <CardFooter className="flex flex-col space-y-4">
+          <CardFooter>
             <Link href={`/vehicles/${vehicle.id}`} className="w-full">
               <Button className="w-full">View Details</Button>
             </Link>
-            <SocialShare
-              url={`https://cryptodrive-.com/vehicles/${vehicle.id}`}
-              title={`Check out this ${vehicle.name} on CryptoDrive !`}
-            />
           </CardFooter>
         </Card>
       ))}
@@ -58,3 +65,4 @@ export default function VehicleGrid() {
   )
 }
 
+export default VehicleGrid
